convert all category commands into slash commands
